requirements.txt
anyascii==0.3.1
asgiref==3.5.0
beautifulsoup4==4.9.3
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.8
django-cogwheels==0.3
django-filter==21.1
django-modelcluster==5.3
django-taggit==2.1.0
django-treebeard==4.5.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
icalendar==4.0.9
idna==3.3
l18n==2021.3
openpyxl==3.0.9
Pillow==9.1.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
Unidecode==1.3.4
urllib3==1.26.9
wagtail==2.16.2
wagtail-color-panel==1.3.1
wagtailgeowidget==6.1.0
wagtailmenus==3.0.2
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
